broke out the tet servers into their own files for each service
