+++ pkg/test/mod-e2b.go
@@ -1,0 +1,44 @@
+package test
+
+import (
+	"log"
+	"net/http"
+	"net/http/httptest"
+	"regexp"
+)
+
+// E2bTestServer creates a test server for emulating the e2b api.
+func (ts *ServerTest) E2bTestServer() *httptest.Server {
+	return httptest.NewUnstartedServer(
+		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			log.Printf(
+				"received a %s request at path %q\n",
+				r.Method,
+				r.URL.Path,
+			)
+
+			// check auth
+			if r.Header.Get("X-API-Key") != GetTestToken() &&
+				r.Header.Get("api-key") != GetTestToken() {
+				w.WriteHeader(http.StatusUnauthorized)
+				return
+			}
+
+			// Handle /path/* routes.
+			// Note: the * is converted to a .* in register handler for proper regex handling
+			for route, handler := range ts.handlers {
+				// Adding ^ and $ to make path matching deterministic since go map iteration isn't ordered
+				pattern, _ := regexp.Compile("^" + route + "$")
+				if pattern.MatchString(r.URL.Path) {
+					handler(w, r)
+					return
+				}
+			}
+			http.Error(
+				w,
+				"the resource path doesn't exist",
+				http.StatusNotFound,
+			)
+		}),
+	)
+}
