+++ pkg/test/server.go
@@ -1,10 +1,7 @@
 package test
 
 import (
-	"log"
 	"net/http"
-	"net/http/httptest"
-	"regexp"
 	"strings"
 )
 
@@ -38,112 +35,4 @@ func (ts *ServerTest) RegisterHandler(path string, handler handler) {
 	// in GroqTestServer
 	path = strings.ReplaceAll(path, "*", ".*")
 	ts.handlers[path] = handler
-}
-
-// GroqTestServer Creates a mocked Groq server which can pretend to handle requests during testing.
-func (ts *ServerTest) GroqTestServer() *httptest.Server {
-	return httptest.NewUnstartedServer(
-		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			log.Printf(
-				"received a %s request at path %q\n",
-				r.Method,
-				r.URL.Path,
-			)
-
-			// check auth
-			if r.Header.Get("Authorization") != "Bearer "+GetTestToken() &&
-				r.Header.Get("api-key") != GetTestToken() {
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-
-			// Handle /path/* routes.
-			// Note: the * is converted to a .* in register handler for proper regex handling
-			for route, handler := range ts.handlers {
-				// Adding ^ and $ to make path matching deterministic since go map iteration isn't ordered
-				pattern, _ := regexp.Compile("^" + route + "$")
-				if pattern.MatchString(r.URL.Path) {
-					handler(w, r)
-					return
-				}
-			}
-			http.Error(
-				w,
-				"the resource path doesn't exist",
-				http.StatusNotFound,
-			)
-		}),
-	)
-}
-
-// E2bTestServer creates a test server for emulating the e2b api.
-func (ts *ServerTest) E2bTestServer() *httptest.Server {
-	return httptest.NewUnstartedServer(
-		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			log.Printf(
-				"received a %s request at path %q\n",
-				r.Method,
-				r.URL.Path,
-			)
-
-			// check auth
-			if r.Header.Get("X-API-Key") != GetTestToken() &&
-				r.Header.Get("api-key") != GetTestToken() {
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-
-			// Handle /path/* routes.
-			// Note: the * is converted to a .* in register handler for proper regex handling
-			for route, handler := range ts.handlers {
-				// Adding ^ and $ to make path matching deterministic since go map iteration isn't ordered
-				pattern, _ := regexp.Compile("^" + route + "$")
-				if pattern.MatchString(r.URL.Path) {
-					handler(w, r)
-					return
-				}
-			}
-			http.Error(
-				w,
-				"the resource path doesn't exist",
-				http.StatusNotFound,
-			)
-		}),
-	)
-}
-
-// ToolhouseTestServer creates a test server for emulating the toolhouse api.
-func (ts *ServerTest) ToolhouseTestServer() *httptest.Server {
-	return httptest.NewUnstartedServer(
-		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			log.Printf(
-				"received a %s request at path %q\n",
-				r.Method,
-				r.URL.Path,
-			)
-
-			// check auth
-			if r.Header.Get("Authorization") != "Bearer "+GetTestToken() &&
-				r.Header.Get("api-key") != GetTestToken() {
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-
-			// Handle /path/* routes.
-			// Note: the * is converted to a .* in register handler for proper regex handling
-			for route, handler := range ts.handlers {
-				// Adding ^ and $ to make path matching deterministic since go map iteration isn't ordered
-				pattern, _ := regexp.Compile("^" + route + "$")
-				if pattern.MatchString(r.URL.Path) {
-					handler(w, r)
-					return
-				}
-			}
-			http.Error(
-				w,
-				"the resource path doesn't exist",
-				http.StatusNotFound,
-			)
-		}),
-	)
 }
